+++ src/Session/Name.php
@@ -14,11 +14,8 @@ use Innmind\Immutable\{
  */
 final class Name
 {
-    private string $value;
-
-    private function __construct(string $value)
+    private function __construct(private string $value)
     {
-        $this->value = $value;
     }
 
     /**
@@ -26,6 +23,7 @@ final class Name
      *
      * @return Maybe<self>
      */
+    #[\NoDiscard]
     public static function maybe(string $value): Maybe
     {
         return Maybe::just(Str::of($value))
@@ -33,6 +31,7 @@ final class Name
             ->map(static fn($value) => new self($value->toString()));
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value;